requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API Clients
amadeus==9.0.0
httpx==0.25.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HubSpot Integration
hubspot-api-client==8.3.0

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Task Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Notifications
slack-sdk==3.26.1
python-dotenv==1.0.0
twilio==8.10.0  # Twilio SMS
clx-sdk-xms==1.0.0  # Sinch SMS (backup)

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing async clients

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Monitoring
sentry-sdk==1.39.1
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
